refactor(desktop): move the Brain Map bridge actions, the notification waiter and the rebuild outcome out of oversized files
The product-file line-count ratchet flagged the base bridge file, the atlas
view model and RewindDatabase. The Brain Map actions now live in
DesktopAutomationBridge+MemoryAtlas.swift (the flow lint and the registry
tests glob every DesktopAutomationBridge*.swift), NotificationWaiter is its
own file, the rebuild outcome type sits with the rebuild in
BrainMapLocalRebuild.swift, and the local_kg_edges citations migration is
registered through LocalKGSchema next to the knowledge-graph store.

Verification: swift test --filter "BrainMapLocalRebuild|MemoryAtlas|MemoryGraph|MemoryHubBrainMap|DesktopAutomationSecondaryAction|GlassContentChrome|KnowledgeGraph|ViewExporterRegistry" (236 tests, 0 failures); pinned swift-format lint clean.
